refactor: address review comments
- Add tracing::warn log when module size limit is exceeded
- Change ModuleSizeLimitExceeded error fields from usize to u64 for cross-platform consistency
- Document edge case when max_module_size is set to 0

+++ crates/runtime/src/errors.rs
@@ -48,7 +48,7 @@ pub enum FunctionCallError {
     #[error("the method call returned an error: {0:?}")]
     ExecutionError(Vec<u8>),
     #[error("module size limit exceeded: size {size} bytes exceeds maximum {max} bytes")]
-    ModuleSizeLimitExceeded { size: usize, max: usize },
+    ModuleSizeLimitExceeded { size: u64, max: u64 },
 }
 
 #[derive(Debug, Serialize, ThisError)]
